requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.0.1
Flask-JWT-Extended==4.2.3
Flask-Bcrypt==0.7.1
Flask-Marshmallow==0.14.0
marshmallow-sqlalchemy==0.26.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0